Extracted BookAChat and ProfileModal components
